changed humidity and pressure value to integer

+++ rpims.py
@@ -324,8 +324,8 @@ def oled_sh1106():
                 #get data from redis db
                 values = redis_db.mget('BME280_Temperature', 'BME280_Humidity', 'BME280_Pressure', 'door_sensor_1', 'door_sensor_2', 'CPU_Temperature')
                 temperature = round(float(values[0]),1)
-                humidity = round(float(values[1]),1)
-                pressure = round(float(values[2]),1)
+                humidity = int(round(float(values[1]),1))
+                pressure = int(round(float(values[2]),1))
                 door_sensor_1 = values[3]
                 door_sensor_2 = values[4]
                 cputemp = round(float(values[5]),1)
@@ -382,8 +382,8 @@ def lcd_st7735():
             #get data from redis db
             values = redis_db.mget('BME280_Temperature', 'BME280_Humidity', 'BME280_Pressure', 'door_sensor_1', 'door_sensor_2', 'CPU_Temperature')
             temperature = round(float(values[0]),1)
-            humidity = round(float(values[1]),1)
-            pressure = round(float(values[2]),1)
+            humidity = int(round(float(values[1]),1))
+            pressure = int(round(float(values[2]),1))
             door_sensor_1 = values[3]
             door_sensor_2 = values[4]
             cputemp = round(float(values[5]),1)
